add CRUD loaisanpham, fix del nhomsanpham related to loaisanpham

+++ app/controllers/admin/AdminDefaultController.php
@@ -103,8 +103,13 @@ class AdminDefaultController extends \AdminController {
 		$this->layout->nest('content', 'admin.admin.products_type', array('types' => $types));
 	}	
 	public function type_destroy($id){
-		Nhomsanpham::destroy($id);
-		return Redirect::route('admin.products_type')->with(Session::flash('notify', "Delete products type: $id successfully"));
+		$items = Loaisanpham::where('id_nhom','=',$id)->count();
+		if($items){
+			return Redirect::route('admin.products_type')->with(Session::flash('notify', "Cannot delete products type: $id because exist product belong to this type"));			
+		}else{
+			Nhomsanpham::destroy($id);
+			return Redirect::route('admin.products_type')->with(Session::flash('notify', "Delete products type: $id successfully"));
+		}
 	}	
 	public function type_new(){
 		$this->layout->title = "Achxi :: Add New Product Type";
@@ -143,11 +148,70 @@ class AdminDefaultController extends \AdminController {
 			return Redirect::route('admin.type_edit_form', $id)->withErrors($validator)->withInput();
 		}else{
 			Nhomsanpham::where('id_nhom','=',$id)->update($data);
+			$count = Loaisanpham::where('id_nhom','=',$id)->count();
 			$items = Loaisanpham::where('id_nhom','=',$id);
-			if($items){
+			if($count){
 				$items->update(array('id_nhom' => Input::get('id_nhom')));
 			}
 			return Redirect::route('admin.products_type')->with(Session::flash('notify', 'Update product type: '.$id.' successfully'));
 		}				
 	}
+	public function products_kind(){
+		$this->layout->title = "Achxi :: Products Kind";
+		$kinds = Loaisanpham::all();
+		$this->layout->nest('content', 'admin.admin.products_kind', array('kinds' => $kinds));
+	}
+	public function kind_destroy($id){
+		Loaisanpham::destroy($id);
+		return Redirect::route('admin.products_kind')->with(Session::flash('notify', "Delete products kind: $id successfully"));
+	}
+	public function kind_new(){
+		$this->layout->title = "Achxi :: Add New Product Kind";
+		$types = Nhomsanpham::all();
+		$this->layout->nest('content', 'admin.admin.kind_new', array('types' => $types));
+	}
+	public function kind_add(){
+		$data = array(  'id_loai' => Input::get('id_loai'),
+						'id_nhom' => Input::get('id_nhom'),
+						'tenloaisp' => Input::get('tenloaisp')
+					);
+
+		$validator = Validator::make($data, Loaisanpham::$add_rules);
+		if($validator->fails()){
+			return Redirect::route('admin.kind_new')->withErrors($validator)->withInput();
+		}else{
+			Loaisanpham::create($data);
+			return Redirect::route('admin.products_kind')->with(Session::flash('notify', 'Create products kind '.Input::get('tennhom'). ' successfully'));
+		}	
+	}
+	public function kind_edit_form($id){
+		$this->layout->title = "Achxi :: Edit Product Kind";
+		$kind = Loaisanpham::find($id);
+		$types = Nhomsanpham::all();
+		$this->layout->nest('content', 'admin.admin.kind_edit_form', array('kind' => $kind,
+																			'types' => $types,
+																			'id' => $id));
+
+	}
+	public function kind_edit(){
+		$id = Input::get('id');
+		$data = array('id_loai' => Input::get('id_loai'),
+					  'id_nhom' => Input::get('id_nhom'),
+					  'tenloaisp' => Input::get('tenloaisp'),
+			);
+
+		$validator = Validator::make(
+			$data, array(
+				'id_loai' 	=> 'required|numeric|min:1|unique:Loaisanpham,id_loai,'.$id.',id_loai',
+				'tennhom'   => 'min:2'
+			)
+		);
+		if($validator->fails()){
+			return Redirect::route('admin.kind_edit_form', $id)->withErrors($validator)->withInput();
+		}else{
+			Loaisanpham::where('id_loai','=',$id)->update($data);
+			return Redirect::route('admin.products_kind')->with(Session::flash('notify', 'Update product kind: '.$id.' successfully'));
+		}				
+	}
+
 }
